These doubles should be ints

+++ Wiimote3Point/Wiimote3Point/Wiimote3P.cs
@@ -73,8 +73,8 @@ namespace Wiimote3Point
             int PIXELS_Y = 768;
             const double FOV_X = Math.PI / 4;
 
-            double pxFromOriginX = (PIXELS_X/2) - pixelcoords.X;
-            double pxFromOriginY = (PIXELS_Y/2) - pixelcoords.Y;
+            int pxFromOriginX = (PIXELS_X/2) - pixelcoords.X;
+            int pxFromOriginY = (PIXELS_Y/2) - pixelcoords.Y;
             double pxFromOriginZ = (1 / Math.Tan(FOV_X/2)) * PIXELS_X;
             Vector<double> unitVector = new DenseVector(new double[] { pxFromOriginX, pxFromOriginY, pxFromOriginZ });
             return unitVector.Normalize(2);
